FInal Submission
we use the EEG signal to recognize human emotion. EEG is a time sequence signal, which needs a model which can process time sequence signal. To pursue a higher accuracy, LSTM stands for the right choice for dealing with EEG signal. LSTM can increase the accuracy of classification because LSTM has a system called cell. The cell contains four gates to help us filter the unnecessary information and keep critical information along with input data